:art: Reorg the views by usage context
