+++ cnxpublishing/views/templates/base.html
